requirements.txt
gunicorn==19.7.1
Flask==0.12.2
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.3.1
Flask-SSLify==0.1.5
psycopg2==2.7.4
